Moving DEV and old TXT to a DEV folder
