+++ src/main/java/frc/robot/DEV/CBA1Input_DEV.java
@@ -1,12 +1,12 @@
-package frc.robot;
+package frc.robot.DEV;
 
-public class CBA1Input {
+public class CBA1Input_DEV {
     
 public double x, y, theta, interval; //joint declaration, usable for seperate types of same decoration and type
 
 
 
-public CBA1Input(double x, double y, double interval) {
+public CBA1Input_DEV(double x, double y, double interval) {
     this.x = x;
     this.y = y;
     this.interval = interval;
@@ -14,7 +14,7 @@ public CBA1Input(double x, double y, double interval) {
   
 }
     
-public CBA1Input(double x, double y,double interval, double theta ) {
+public CBA1Input_DEV(double x, double y,double interval, double theta ) {
     this.x = x;
     this.y = y;
     //Allows us to implement rotations later on
