+++ src/main/java/frc/robot/DEV/SemiAuto_02.java
@@ -5,9 +5,10 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.Launcher;
+package frc.robot.DEV;
 
 import frc.robot.commands.IndexSpeed;
+import frc.robot.commands.Launcher.LaunchSpeed;
 import frc.robot.subsystems.Index;
 import frc.robot.subsystems.Launcher;
