+++ src/main/java/frc/robot/DEV/SemiAuto_09.java
@@ -4,7 +4,7 @@
 // the WPILib BSD license file in the root directory of this project.
 
 // FRC 3603 Launch High
-package frc.robot.commands.Launcher;
+package frc.robot.DEV;
 
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Launcher;
 
 
-public class SemiAuto_10 extends CommandBase {
+public class SemiAuto_09 extends CommandBase {
     private final Launcher launcher;
   
     // https://docs.limelightvision.io/en/latest/cs_drive_to_goal_2019.html
@@ -20,9 +20,18 @@ public class SemiAuto_10 extends CommandBase {
     static boolean targetStatus = false;
     NetworkTable limelight = NetworkTableInstance.getDefault().getTable("limelight-back");//Instantiate the tables
     double tx = limelight.getEntry("tx").getDouble(0.0);
-    double ty = limelight.getEntry("ty").getDouble(0.0);
-
-  public SemiAuto_10(Launcher launch) {
+      
+    public static boolean getTargetStatus(double tx) {
+      if (tx < 5) {
+        System.out.println("Missile lock-on");
+        targetStatus = true; // Don't know if this is needed if call 'getTargetStatus'
+        return true;
+      } else {
+        System.out.println("Still seeking target");
+        targetStatus = false; // Don't know if this is needed if call 'getTargetStatus'
+        return false;}
+     }  
+  public SemiAuto_09(Launcher launch) {
 
     launcher = launch;
     addRequirements(launcher);
@@ -35,18 +44,7 @@ public class SemiAuto_10 extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    if (-5 < tx && tx < 5) {
-      // launcher.setLaunch2();
-      System.out.println("Missile lock-on");
-      System.out.println(tx);
-      System.out.println(ty);
-    } else {
-      System.out.println("Still seeking target");
-      System.out.println(tx);
-      System.out.println(ty);
-      // launcher.stopLauncher();
-   }  
-    
+    launcher.setLaunch2();
   }
 
   // Called once the command ends or is interrupted.
