+++ src/main/java/frc/robot/DEV/SemiAuto_15.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.Launcher;
+package frc.robot.DEV;
 
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
@@ -11,12 +11,10 @@ import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.VisionRange;
-// import frc.robot.Constants.Launcher;
 import frc.robot.subsystems.Launcher;
 
 
-public class LaunchSemiAuto extends CommandBase {
+public class SemiAuto_15 extends CommandBase {
 
   private final Launcher launcher;
   boolean targetLock = false;
@@ -35,10 +33,6 @@ public class LaunchSemiAuto extends CommandBase {
   NetworkTableEntry ty = table.getEntry("ty");
   NetworkTableEntry ta = table.getEntry("ta");
 
-  double TX = tx.getDouble(0.0);
-  double TY = ty.getDouble(0.0);
-  double area = ta.getDouble(0.0);
-
   ShuffleboardTab visionTab = Shuffleboard.getTab("Vision");
 
   /** 
@@ -48,7 +42,7 @@ public class LaunchSemiAuto extends CommandBase {
   */
 
   
-  public LaunchSemiAuto(Launcher launch) {
+  public SemiAuto_15(Launcher launch) {
     // Use addRequirements() here to declare subsystem dependencies.
     launcher = launch;
     addRequirements(launcher);
@@ -66,22 +60,21 @@ public class LaunchSemiAuto extends CommandBase {
     double TY = ty.getDouble(0.0);
     double area = ta.getDouble(0.0);
 
-    SmartDashboard.putNumber(("tX"), TX);
-    SmartDashboard.putNumber(("tY"), TY);
+    SmartDashboard.putNumber(("tX"), TX); // Added since v.12
+    SmartDashboard.putNumber(("tY"), TY); // Added since v.12
     SmartDashboard.putBoolean("Target Status", targetLock);
 
-    if ((VisionRange.txMin< TX && TX < VisionRange.txMax) & (VisionRange.tyMin< TY && TY < VisionRange.tyMax)) {
+    if((5 < TY && TY < 14) & (-5 < TX && TX <5))
+    {
       //Sets targetLock to true when tx & ty are within the parameters
-      System.out.println("++ Target LOCKED ++ " + "(" + TX + "," + TY + ")" + " Area:" + area);
-
-      // TODO Replace with Group2BallsHigh once intake is fixed
+      System.out.println("++ Target LOCKED ++ " + "(" + TX + "," + TY + ")" + " Area:" + area); // Revised since v.12
       launcher.setLaunch2();
-      targetLock = true;
+      targetLock = true; // Added since v.12
 
     } else {
-      System.out.println("-- Target NOT locked -- " + "(" + TX + "," + TY + ")" + " Area:" + area);
+      System.out.println("-- Target NOT locked --" + "(" + TX + "," + TY + ")" + " Area:" + area);
       launcher.stopLauncher();
-      targetLock = false;
+      targetLock = false; // Added since v.12
     } 
   
   }
@@ -89,7 +82,6 @@ public class LaunchSemiAuto extends CommandBase {
   @Override
   public void end(boolean interrupted) {      
       launcher.stopLauncher();}
-      // FIXME Do TX and TY need to be reset to zero to prevent false positives with next attempt at a shot?
       // double TX = tx.getDouble(0.0);
       // double TY = ty.getDouble(0.0);
       // double area = ta.getDouble(0.0);
