+++ src/main/java/frc/robot/DEV/SemiAuto_16.java
@@ -2,22 +2,22 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.Launcher;
+package frc.robot.DEV;
 
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.Constants.VisionRange;
+// import frc.robot.Constants.Launcher;
 import frc.robot.subsystems.Launcher;
 
 
-public class SemiAuto_15 extends CommandBase {
+public class SemiAuto_16 extends CommandBase {
 
   private final Launcher launcher;
-  boolean targetLock = false;
 
 
   //Shuffleboard.selectTab("Vision");
@@ -33,6 +33,10 @@ public class SemiAuto_15 extends CommandBase {
   NetworkTableEntry ty = table.getEntry("ty");
   NetworkTableEntry ta = table.getEntry("ta");
 
+  double TX = tx.getDouble(0.0);
+  double TY = ty.getDouble(0.0);
+  double area = ta.getDouble(0.0);
+
   ShuffleboardTab visionTab = Shuffleboard.getTab("Vision");
 
   /** 
@@ -42,7 +46,7 @@ public class SemiAuto_15 extends CommandBase {
   */
 
   
-  public SemiAuto_15(Launcher launch) {
+  public SemiAuto_16(Launcher launch) {
     // Use addRequirements() here to declare subsystem dependencies.
     launcher = launch;
     addRequirements(launcher);
@@ -58,33 +62,21 @@ public class SemiAuto_15 extends CommandBase {
     // read values periodically
     double TX = tx.getDouble(0.0);
     double TY = ty.getDouble(0.0);
-    double area = ta.getDouble(0.0);
 
-    SmartDashboard.putNumber(("tX"), TX); // Added since v.12
-    SmartDashboard.putNumber(("tY"), TY); // Added since v.12
-    SmartDashboard.putBoolean("Target Status", targetLock);
-
-    if((5 < TY && TY < 14) & (-5 < TX && TX <5))
-    {
+    if ((VisionRange.txMin< TX && TX < VisionRange.txMax) && (VisionRange.tyMin< TY && TY < VisionRange.tyMax)) {
       //Sets targetLock to true when tx & ty are within the parameters
-      System.out.println("++ Target LOCKED ++ " + "(" + TX + "," + TY + ")" + " Area:" + area); // Revised since v.12
       launcher.setLaunch2();
-      targetLock = true; // Added since v.12
 
     } else {
-      System.out.println("-- Target NOT locked --" + "(" + TX + "," + TY + ")" + " Area:" + area);
       launcher.stopLauncher();
-      targetLock = false; // Added since v.12
     } 
   
   }
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {      
-      launcher.stopLauncher();}
-      // double TX = tx.getDouble(0.0);
-      // double TY = ty.getDouble(0.0);
-      // double area = ta.getDouble(0.0);
+      launcher.stopLauncher();
+  }
 
   // Returns true when the command should end.
   @Override
