+++ src/main/java/frc/robot/Robot.java
@@ -4,8 +4,6 @@
 
 package frc.robot;
 
-import java.lang.annotation.Target;
-
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -20,9 +18,6 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.Constants.VisionRange;
-import frc.robot.subsystems.TargetStatus;
-
-
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -97,6 +92,7 @@ public class Robot extends TimedRobot {
     
     SmartDashboard.putNumber(("tX"), TX);
     SmartDashboard.putNumber(("tY"), TY);
+    SmartDashboard.putNumber("Area", area);
     SmartDashboard.putBoolean("Target Status", targetLock);
 
   }
