+++ src/main/java/frc/robot/commands/Launcher/SemiAutoLaunch.java
@@ -6,19 +6,18 @@ package frc.robot.commands.Launcher;
 
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.Constants.VisionRange;
 // import frc.robot.Constants.Launcher;
 import frc.robot.subsystems.Launcher;
 
 
-public class SemiAuto_14 extends CommandBase {
+public class SemiAutoLaunch extends CommandBase {
 
   private final Launcher launcher;
-  boolean targetLock = false;
 
 
   //Shuffleboard.selectTab("Vision");
@@ -34,6 +33,10 @@ public class SemiAuto_14 extends CommandBase {
   NetworkTableEntry ty = table.getEntry("ty");
   NetworkTableEntry ta = table.getEntry("ta");
 
+  double TX = tx.getDouble(0.0);
+  double TY = ty.getDouble(0.0);
+  double area = ta.getDouble(0.0);
+
   ShuffleboardTab visionTab = Shuffleboard.getTab("Vision");
 
   /** 
@@ -43,7 +46,7 @@ public class SemiAuto_14 extends CommandBase {
   */
 
   
-  public SemiAuto_14(Launcher launch) {
+  public SemiAutoLaunch(Launcher launch) {
     // Use addRequirements() here to declare subsystem dependencies.
     launcher = launch;
     addRequirements(launcher);
@@ -59,19 +62,12 @@ public class SemiAuto_14 extends CommandBase {
     // read values periodically
     double TX = tx.getDouble(0.0);
     double TY = ty.getDouble(0.0);
-    double area = ta.getDouble(0.0);
 
-    SmartDashboard.putNumber(("tX"), TX); // Added since v.12
-    SmartDashboard.putNumber(("tY"), TY); // Added since v.12
-
-    if((5 < TY && TY < 14) & (-5 < TX && TX <5))
-    {
+    if ((VisionRange.txMin< TX && TX < VisionRange.txMax) && (VisionRange.tyMin< TY && TY < VisionRange.tyMax)) {
       //Sets targetLock to true when tx & ty are within the parameters
-      System.out.println("++ Target LOCKED ++ " + "(" + TX + "," + TY + ")" + " Area:" + area); // Revised since v.12
       launcher.setLaunch2();
 
     } else {
-      System.out.println("-- Target NOT locked --" + "(" + TX + "," + TY + ")" + " Area:" + area);
       launcher.stopLauncher();
     } 
   
@@ -79,7 +75,8 @@ public class SemiAuto_14 extends CommandBase {
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {      
-      launcher.stopLauncher();}
+      launcher.stopLauncher();
+  }
 
   // Returns true when the command should end.
   @Override
